dns-explorer-backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
dnspython==2.6.1
pydantic==2.8.2
python-dotenv==1.0.1
